rm the original BeamBoundedSourceVertex
